LayoutTests/fast/animation/request-animation-frame-display-expected.txt

Tests requestAnimationFrame callback handling of display: property changed within another callback

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS callbackInvokedOnA is true
PASS successfullyParsed is true

TEST COMPLETE
